여정리스트 get요청 url 수정
