+++ src/app/email/signup/nicknameCheck/page.tsx
